Put GithubManager classes into their own files
